Installation restructurization
The main idea was to combine everything related to installation to the
corresponding section but separated by subsections depending on user
level.

Also the question is if we need a "Building from the source" in
"Installation" as after adding Homebrew formula 90% users will never
need it, but the last 10% will build it from source when hacking it,
i.e. it sounds more like instruction for contribution, which can live in
a separate Contribute.MD file.